refactor(front): eliminate legacy pages/ layer, route directly to features
- moved feature-owned styles out of pages/ into their features:
  pages/tables/TablesPage.styles -> features/tables/ui/
  pages/sima/SimaTablesPage.styles -> features/sima/tables/ui/
  (removes 13 backward feature->pages imports)
- moved the real SimaRecordsListPage component into features/sima/records/
- deleted all pages/ re-export shims; App.tsx now imports feature entries directly
