+++ front/src/features/sima/records/SimaRecordsListPage.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import styled from "styled-components";
 import { useQuery } from "@tanstack/react-query";
-import { getSima } from "../../api/simaApi";
-import SimaTable from "../../components/SimaTable";
-import type { Sima } from "../../types/sima";
+import { getSima } from "../../../api/simaApi";
+import SimaTable from "../../../components/SimaTable";
+import type { Sima } from "../../../types/sima";
 
 const PageContainer = styled.div`
   flex: 1;
