+++ front/src/features/tables/ui/TablesPage.styles.tsx
@@ -1,6 +1,6 @@
-import { JSX } from "react";
-import styled, { keyframes } from "styled-components";
-import { lightenHex } from "../../utils/limnologicData";
+/* prettier-ignore-file */
+import styled from "styled-components";
+import { lightenHex } from "../../../utils/limnologicData";
 
 export const PRIMARY_BLUE = "#0b5fff";
 export const PRIMARY_BLUE_HOVER = "#2a7bff";
@@ -8,6 +8,7 @@ export const MUTED_BLUE = "#e8f1ff";
 export const TEXT_DARK = "#0b2740";
 export const SURFACE = "#ffffff";
 export const BORDER = "#e6eefb";
+export const LEGEND_BG = "#f8fbff";
 
 export const Page = styled.div`
   min-height: 100vh;
@@ -43,7 +44,7 @@ export const LeftColumn = styled.div`
 `;
 
 export const Controls = styled.div`
-  background: #fff;
+  background: ${SURFACE};
   padding: 16px;
   border-radius: 12px;
   box-shadow: 0 8px 28px rgba(6, 58, 128, 0.06);
@@ -57,12 +58,21 @@ export const Row = styled.div`
   display: flex;
   gap: 10px;
   align-items: center;
+
+  @media (max-width: 520px) {
+    flex-direction: column;
+    align-items: stretch;
+  }
 `;
 
 export const Label = styled.label`
   font-size: 13px;
   color: #334155;
   min-width: 100px;
+
+  @media (max-width: 520px) {
+    min-width: auto;
+  }
 `;
 
 export const Select = styled.select`
@@ -96,6 +106,18 @@ export const ColumnItem = styled.label`
   }
 `;
 
+export const ColumnsBox = styled.div`
+  background: ${SURFACE};
+  padding: 14px;
+  border-radius: 12px;
+  box-shadow: 0 8px 28px rgba(6, 58, 128, 0.04);
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  overflow: auto;
+  flex: 1 1 auto;
+`;
+
 export const RightPanel = styled.div`
   display: flex;
   flex-direction: column;
@@ -114,47 +136,21 @@ export const ControlsTopRight = styled.div`
   }
 `;
 
-export const Button = styled.button<{ $primary?: boolean }>`
-  padding: 10px 14px;
-  border-radius: 10px;
-  border: ${(p) => (p.$primary ? "none" : `1px solid ${BORDER}`)};
-  cursor: pointer;
-  font-weight: 700;
-  background: ${(p) => (p.$primary ? PRIMARY_BLUE : SURFACE)};
-  color: ${(p) => (p.$primary ? "#fff" : TEXT_DARK)};
-  display: inline-flex;
+export const DownloadButtonsContainer = styled.div`
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
   align-items: center;
-  gap: 8px;
-  box-shadow: ${(p) => (p.$primary ? "0 6px 18px rgba(11,95,255,0.12)" : "none")};
-  transition:
-    transform 0.12s ease,
-    background 0.12s ease,
-    box-shadow 0.12s ease;
-  border-radius: 8px;
-
-  &:hover {
-    transform: translateY(-2px);
-    background: ${(p) => (p.$primary ? PRIMARY_BLUE_HOVER : MUTED_BLUE)};
-  }
-
-  &:disabled {
-    opacity: 0.6;
-    cursor: not-allowed;
-    transform: none;
-    box-shadow: none;
-  }
-
-  @media (max-width: 520px) {
-    width: 100%;
-  }
 `;
 
 export const Panel = styled.div`
-  background: #fff;
+  background: ${SURFACE};
   padding: 20px;
   border-radius: 12px;
   height: 100%;
   box-shadow: 0 12px 36px rgba(9, 30, 66, 0.06);
+  display: flex;
+  flex-direction: column;
 `;
 
 export const ChartWrapper = styled.div`
@@ -169,52 +165,6 @@ export const ChartMain = styled.div`
   min-height: 420px;
 `;
 
-export const Tooltip = styled.div<{ left: number; top: number; color?: string }>`
-  position: absolute;
-  left: ${(p) => p.left}px;
-  top: ${(p) => p.top}px;
-  transform: translate(-8px, -100%);
-  background: #ffffff;
-  border-radius: 8px;
-  padding: 8px 10px;
-  box-shadow: 0 6px 20px rgba(2, 6, 23, 0.12);
-  font-size: 13px;
-  min-width: 160px;
-  max-width: 340px;
-  max-height: 320px;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-`;
-
-export const TooltipStationList = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-  max-height: 220px;
-  overflow: hidden;
-`;
-
-export const TooltipFooter = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-export const Legend = styled.div`
-  display: flex;
-  gap: 12px;
-  flex-wrap: wrap;
-  margin-top: 12px;
-`;
-
-export const LegendItem = styled.div`
-  display: flex;
-  gap: 8px;
-  align-items: center;
-`;
-
 export const MapPlaceholder = styled.div`
   position: relative;
   flex: 1;
@@ -222,6 +172,17 @@ export const MapPlaceholder = styled.div`
   border-radius: 8px;
   overflow: hidden;
   min-height: 260px;
+  display: flex;
+  flex-direction: column;
+`;
+
+export const MapInner = styled.div`
+  flex: 1;
+  width: 100%;
+  display: flex;
+  align-items: stretch;
+  justify-content: stretch;
+  position: relative;
 `;
 
 export const ZoomControls = styled.div`
@@ -245,13 +206,28 @@ export const ZoomControls = styled.div`
     cursor: pointer;
     font-weight: 700;
   }
+`;
 
-  label {
-    color: #e6f0ff;
-    font-size: 13px;
-    display: flex;
-    align-items: center;
-    gap: 8px;
+export const SimaTableWrapper = styled.div`
+  table thead th {
+    background: linear-gradient(180deg, ${PRIMARY_BLUE} 0%, ${PRIMARY_BLUE_HOVER} 100%) !important;
+    color: #fff !important;
+    font-weight: 700 !important;
+    position: sticky !important;
+    top: 0 !important;
+    z-index: 3 !important;
+    font-size: 13px !important;
+    padding: 10px 12px !important;
+  }
+
+  table tbody tr:hover {
+    background: ${MUTED_BLUE} !important;
+  }
+
+  table tbody td {
+    padding: 10px 12px !important;
+    color: ${TEXT_DARK} !important;
+    font-size: 14px !important;
   }
 `;
 
@@ -293,43 +269,6 @@ export const TableElement = styled.table`
   }
 `;
 
-const spin = keyframes`
-  0% { transform: rotate(0deg) }
-  100% { transform: rotate(360deg) }
-`;
-
-const SpinnerWrapper = styled.div`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 24px;
-  height: 24px;
-`;
-
-const SpinnerCircle = styled.div`
-  width: 18px;
-  height: 18px;
-  border-radius: 50%;
-  border: 3px solid rgba(37, 99, 235, 0.15);
-  border-top-color: rgba(37, 99, 235, 1);
-  animation: ${spin} 0.9s linear infinite;
-`;
-
-export function Spinner(): JSX.Element {
-  return (
-    <SpinnerWrapper aria-hidden>
-      <SpinnerCircle />
-    </SpinnerWrapper>
-  );
-}
-
-export const DownloadButtonsContainer = styled.div`
-  display: flex;
-  gap: 10px;
-  flex-wrap: no-wrap;
-  align-items: center;
-`;
-
 export const DownloadButton = styled.button<{ variant: "csv" | "json" | "pdf" }>`
   padding: 8px 14px;
   border: none;
@@ -367,4 +306,68 @@ export const DownloadButton = styled.button<{ variant: "csv" | "json" | "pdf" }>
       : p.variant === "json"
         ? `border-left: 4px solid ${PRIMARY_BLUE_HOVER};`
         : `border-left: 4px solid ${lightenHex(PRIMARY_BLUE, -0.05)};`}
+`;
+
+export const Button = styled.button<{ $primary?: boolean }>`
+  padding: 10px 14px;
+  border-radius: 10px;
+  border: ${(p) => (p.$primary ? "none" : `1px solid ${BORDER}`)};
+  cursor: pointer;
+  font-weight: 700;
+  background: ${(p) => (p.$primary ? PRIMARY_BLUE : SURFACE)};
+  color: ${(p) => (p.$primary ? "#fff" : TEXT_DARK)};
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  box-shadow: ${(p) => (p.$primary ? "0 6px 18px rgba(11,95,255,0.12)" : "none")};
+  transition:
+    transform 0.12s ease,
+    background 0.12s ease,
+    box-shadow 0.12s ease;
+  border-radius: 8px;
+
+  &:hover {
+    transform: translateY(-2px);
+    background: ${(p) => (p.$primary ? PRIMARY_BLUE_HOVER : MUTED_BLUE)};
+  }
+
+  &:disabled {
+    opacity: 0.6;
+    cursor: not-allowed;
+    transform: none;
+    box-shadow: none;
+  }
+`;
+
+export const LegendBox = styled.div`
+  margin-top: 12px;
+  padding: 10px;
+  background: ${LEGEND_BG};
+  border-radius: 8px;
+  border: 1px solid rgba(255, 255, 255, 0.04);
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  align-items: center;
+`;
+
+export const LegendItem = styled.div`
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  padding: 6px 8px;
+  border-radius: 6px;
+  background: white;
+  box-shadow: 0 6px 16px rgba(6, 58, 128, 0.04);
+  font-size: 13px;
+  color: ${TEXT_DARK};
+`;
+
+export const Swatch = styled.span<{ color: string }>`
+  width: 14px;
+  height: 14px;
+  border-radius: 50%;
+  display: inline-block;
+  background: ${(p) => p.color};
+  border: 1px solid rgba(0, 0, 0, 0.06);
 `;
