--- front/src/pages/HomePage.tsx
@@ -1,1 +1,0 @@
-export { default } from "../features/home/HomePage";
