--- front/src/pages/SimaPage.tsx
@@ -1,1 +1,0 @@
-export { default } from "./sima/SimaRecordsListPage";
