--- front/src/pages/SimaTablesPage.tsx
@@ -1,1 +1,0 @@
-export { default } from "./sima/SimaTablesPage";
