--- front/src/pages/SimasPage.tsx
@@ -1,1 +1,0 @@
-export { default } from "./sima/SimaInstitutionalPage";
