--- front/src/pages/TablesPage.tsx
@@ -1,1 +1,0 @@
-export { default } from "./tables/TablesPage";
