--- front/src/pages/sima/SimaInstitutionalPage.tsx
@@ -1,1 +1,0 @@
-export { default } from "../../features/sima/institutional/InstitutionalPage";
