--- front/src/pages/sima/SimaTablesPage.tsx
@@ -1,1 +1,0 @@
-export { default } from "../../features/sima/tables/SimaTablesPage";
