--- front/src/pages/tables/TablesPage.tsx
@@ -1,1 +1,0 @@
-export { default } from "../../features/tables/TablesPage";
